+++ StudioGhilbiApps/Modules/MovieList/Model/Movie.swift
