--- StudioGhilbiApps/MovileCollectionViewCell.xib
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="" id="gTV-IL-0wX" customClass="MovieCollectionViewCell" customModule="StudioGhilbiApps" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="376" height="200"/>
-            <autoresizingMask key="autoresizingMask"/>
-            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
-                <rect key="frame" x="0.0" y="0.0" width="376" height="200"/>
-                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                <subviews>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="Movie Title" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gvH-dd-ltf">
-                        <rect key="frame" x="170" y="10" width="186" height="22"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <fontDescription key="fontDescription" name="Arial-BoldMT" family="Arial" pointSize="17"/>
-                        <color key="textColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="Year" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Rxe-SD-FCk">
-                        <rect key="frame" x="170" y="45" width="42" height="15"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <fontDescription key="fontDescription" name="ArialMT" family="Arial" pointSize="13"/>
-                        <color key="textColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" fixedFrame="YES" scrollEnabled="NO" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" editable="NO" text="Movie Descriptions" textAlignment="natural" selectable="NO" translatesAutoresizingMaskIntoConstraints="NO" id="WQX-r5-KPw">
-                        <rect key="frame" x="170" y="82" width="186" height="108"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <color key="textColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <fontDescription key="fontDescription" name="ArialMT" family="Arial" pointSize="10"/>
-                        <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
-                    </textView>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="qKP-WM-kq8">
-                        <rect key="frame" x="12" y="10" width="120" height="180"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="120" id="1Hf-Hq-kXk"/>
-                            <constraint firstAttribute="width" secondItem="qKP-WM-kq8" secondAttribute="height" multiplier="4:6" id="qtT-0O-8im"/>
-                        </constraints>
-                    </imageView>
-                </subviews>
-            </view>
-            <constraints>
-                <constraint firstItem="qKP-WM-kq8" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" constant="10" id="HVc-IT-xKr"/>
-                <constraint firstItem="qKP-WM-kq8" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" constant="12" id="yfz-tA-Wy7"/>
-            </constraints>
-            <size key="customSize" width="376" height="260"/>
-            <connections>
-                <outlet property="descriptionTextView" destination="WQX-r5-KPw" id="bg9-6u-dff"/>
-                <outlet property="imageView" destination="qKP-WM-kq8" id="j6M-xV-d4n"/>
-                <outlet property="movieTitleLabel" destination="gvH-dd-ltf" id="DZl-m2-0Nk"/>
-                <outlet property="yearLabel" destination="Rxe-SD-FCk" id="Y5a-P0-day"/>
-            </connections>
-            <point key="canvasLocation" x="297.10144927536237" y="194.86607142857142"/>
-        </collectionViewCell>
-    </objects>
-</document>
