--- StudioGhilbiApps/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  StudioGhilbiApps
-//
-//  Created by Septian-GLI on 04/08/21.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
